Creating a bootstrap archive now includes hab, et al.
Now we include hab, hab-sup, the exporters, etc. Basically anything in
the habitat repo that has a plan associated with it is now pulled in.

Additionally, upload-archive has been split out into a separate command
which is run automatically from create-archive and the final archive
file is not deleted by default. There is an additional environment
variable that can be set if you don't want the final archive file
to be deleted. I found that if there were problems during the upload,
I wanted that file to stick around so I didn't have to re-generate it
from scratch.

Signed-off-by: Josh Black <raskchanky@gmail.com>