Allow for a different origin to be used to pull packages for
I wanted to be able to do development on builder for on-prem without
having rights to the habitat origin.  This PR justs adds the
ability to specify a different origin in bldr.env.

It also sets a default in provision.sh so that existing users
don't run into problems using an existing `bldr.env` file

Signed-off-by: James Casey <james@chef.io>